+++ src/editors/index.ts
@@ -1,4 +1,4 @@
 export { default as SimpleTextEditor } from './SimpleTextEditor';
 export { default as EditorPortal } from './EditorPortal';
 export { default as OldEditorContainer } from './OldEditorContainer';
-export { default as EditorContainer } from './Editor2Container';
+export { default as EditorContainer } from './EditorContainer';
